Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 int main() {
-    float a,Area;
+    int a,Area;
     scanf("&f",&a);
     area=3.14*a*a;
     printf("Area: %.2f",area);
